adjusting backend for new doctypeprovision entity
