+++ backend/src/admin/admin.module.ts
@@ -7,9 +7,10 @@ import { AdminController } from './admin.controller';
 import { AuthenticationModule } from 'src/authentication/authentication.module';
 import { DocumentTemplateModule } from 'src/document_template/document_template.module';
 import { ProvisionModule } from 'src/provision/provision.module';
+import { DocumentTypeModule } from 'src/document_type/document_type.module';
 
 @Module({
-  imports: [HttpModule, AuthenticationModule, DocumentTemplateModule, ProvisionModule],
+  imports: [HttpModule, AuthenticationModule, DocumentTemplateModule, ProvisionModule, DocumentTypeModule],
   providers: [AdminGuard, AdminService],
   exports: [AdminService],
   controllers: [AdminController],
