+++ backend/src/document_data/entities/document_data_provision.entity.ts
@@ -1,7 +1,7 @@
 import { Provision } from 'src/provision/entities/provision.entity';
 import { Entity, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
 import { DocumentData } from './document_data.entity';
-import { DocumentTypeProvision } from 'src/provision/entities/document_type_provision';
+import { DocumentTypeProvision } from 'src/document_type/entities/document_type_provision';
 
 @Entity()
 export class DocumentDataProvision {
