+++ backend/src/document_type/document_type.controller.ts
@@ -43,4 +43,32 @@ export class DocumentTypeController {
   findAll() {
     return this.documentTypeService.findAll();
   }
+
+  @Get('get-group-max/:document_type_id')
+  getGroupMax(@Param('document_type_id') document_type_id: number) {
+    return this.documentTypeService.getGroupMaxByDocTypeId(document_type_id);
+  }
+
+  @Get('get-all-mandatory-provisions/:id')
+  getMandatoryProvisions() {
+    return this.documentTypeService.getMandatoryProvisions();
+  }
+
+  @Get('get-manage-doc-type-provisions/:document_type_id')
+  getManageDocTypeProvisions(@Param('document_type_id') document_type_id: number) {
+    return this.documentTypeService.getManageDocTypeProvisions(document_type_id);
+  }
+
+  @Get('associate-doc-type/:provision_id/:document_type_id')
+  associateDocType(@Param('provision_id') provision_id: number, @Param('document_type_id') document_type_id: number) {
+    return this.documentTypeService.associateDocType(provision_id, document_type_id);
+  }
+
+  @Get('disassociate-doc-type/:provision_id/:document_type_id')
+  disassociateDocType(
+    @Param('provision_id') provision_id: number,
+    @Param('document_type_id') document_type_id: number
+  ) {
+    return this.documentTypeService.disassociateDocType(+provision_id, document_type_id);
+  }
 }
