+++ backend/src/main.ts
@@ -20,8 +20,8 @@ async function bootstrap() {
   const document = SwaggerModule.createDocument(app, config);
   SwaggerModule.setup('api', app, document);
   const appService = app.get(AppService);
-  await appService.initializeDb();
+  // await appService.initializeDb();
 
-  await app.listen(3001);
+  await app.listen(process.env.ticdi_environment == 'DEVELOPMENT' ? 3001 : 3000);
 }
 bootstrap();
