+++ backend/src/provision/dto/provision.dto.ts
@@ -1,9 +1,9 @@
 export class ProvisionDto {
   dtid?: number;
   type?: string;
-  provision_group?: number;
-  provision_group_text?: string;
-  max?: number;
+  // provision_group?: number;
+  // provision_group_text?: string;
+  // max?: number;
   provision_name?: string;
   free_text?: string;
   help_text?: string;
