+++ backend/src/provision/dto/update-provision.dto.ts
@@ -4,9 +4,9 @@ import { ProvisionDto } from './provision.dto';
 export class UpdateProvisionDto extends PickType(ProvisionDto, [
   'dtid',
   'type',
-  'provision_group',
-  'provision_group_text',
-  'max',
+  // 'provision_group',
+  // 'provision_group_text',
+  // 'max',
   'provision_name',
   'free_text',
   'help_text',
