+++ backend/src/provision/provision.controller.ts
@@ -83,37 +83,4 @@ export class ProvisionController {
   disableProvision(@Param('id') id: number) {
     return this.provisionService.disable(id);
   }
-
-  @Get('get-group-max/:document_type_id')
-  getGroupMax(@Param('document_type_id') document_type_id: number) {
-    return this.provisionService.getGroupMaxByDocTypeId(document_type_id);
-  }
-
-  @Get('get-all-mandatory-provisions/:id')
-  getMandatoryProvisions() {
-    return this.provisionService.getMandatoryProvisions();
-  }
-
-  @Get('remove/:id')
-  remove(@Param('id') id: number) {
-    return this.provisionService.remove(id);
-  }
-
-  @Get('get-manage-doc-type-provisions/:document_type_id')
-  getManageDocTypeProvisions(@Param('document_type_id') document_type_id: number) {
-    return this.provisionService.getManageDocTypeProvisions(document_type_id);
-  }
-
-  @Get('associate-doc-type/:provision_id/:document_type_id')
-  associateDocType(@Param('provision_id') provision_id: number, @Param('document_type_id') document_type_id: number) {
-    return this.provisionService.associateDocType(provision_id, document_type_id);
-  }
-
-  @Get('disassociate-doc-type/:provision_id/:document_type_id')
-  disassociateDocType(
-    @Param('provision_id') provision_id: number,
-    @Param('document_type_id') document_type_id: number
-  ) {
-    return this.provisionService.disassociateDocType(+provision_id, document_type_id);
-  }
 }
