+++ backend/src/provision/provision.module.ts
@@ -1,19 +1,12 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Provision } from './entities/provision.entity';
-import { ProvisionGroup } from './entities/provision_group.entity';
 import { ProvisionController } from './provision.controller';
 import { ProvisionService } from './provision.service';
 import { ProvisionVariable } from './entities/provision_variable.entity';
-import { DocumentTypeModule } from 'src/document_type/document_type.module';
 
 @Module({
-  imports: [
-    TypeOrmModule.forFeature([Provision]),
-    TypeOrmModule.forFeature([ProvisionGroup]),
-    TypeOrmModule.forFeature([ProvisionVariable]),
-    DocumentTypeModule,
-  ],
+  imports: [TypeOrmModule.forFeature([Provision]), TypeOrmModule.forFeature([ProvisionVariable])],
   controllers: [ProvisionController],
   providers: [ProvisionService],
   exports: [ProvisionService],
